asdac-requirements.txt
more_itertools
regex
pytest
colorama
mypy
attrs
